(Part of) authorization and user functions

+++ advisor-backend/src/assessment/assessment.controller.ts
@@ -7,7 +7,6 @@ import {
   Param,
   Delete,
   ParseIntPipe,
-  UseGuards,
   ForbiddenException,
   Query,
   NotFoundException,
@@ -27,8 +26,6 @@ import { UpdateAssessmentDto } from './dto/update-assessment.dto';
 import { AssessmentDto } from './dto/assessment.dto';
 import { FeedbackDto } from './dto/feedback.dto';
 import { SaveCheckpointDto } from './dto/save-checkpoint.dto';
-import { RolesGuard } from '../common/guards/roles.guard';
-import { AuthGuard } from '@nestjs/passport';
 import { Roles } from '../common/decorators/roles.decorator';
 import { Role, User } from '@prisma/client';
 import AuthUser from '../common/decorators/auth-user.decorator';
@@ -50,7 +47,7 @@ export class AssessmentController {
    * @param createAssessmentDto Assessment information
    * @returns created assessment
    */
-  @Post()
+  @Post('')
   @ApiConflictResponse({
     description: 'Assessment with this name and type already exists',
   })
@@ -60,7 +57,6 @@ export class AssessmentController {
   @ApiBadRequestResponse({
     description: 'No active templates found',
   })
-  @UseGuards(AuthGuard('jwt'))
   create(
     @Body() createAssessmentDto: CreateAssessmentDto,
     @AuthUser() user: User
@@ -78,6 +74,7 @@ export class AssessmentController {
     type: AssessmentDto,
     isArray: true,
   })
+  @Roles(Role.ADMIN)
   findAll() {
     return this.assessmentService.findAll();
   }
@@ -88,7 +85,6 @@ export class AssessmentController {
    * @returns AssessmentDto[] List of all assessments
    */
   @Get('my')
-  @UseGuards(AuthGuard('jwt'))
   @ApiResponse({
     description: 'Found assessment of USER',
     type: AssessmentDto,
@@ -106,7 +102,16 @@ export class AssessmentController {
   @Get(':assessment_id')
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
-  findOne(@Param('assessment_id', ParseIntPipe) id: number) {
+  async findOne(
+    @Param('assessment_id', ParseIntPipe) id: number,
+    @AuthUser() user: User
+  ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.findOne(id);
   }
 
@@ -122,10 +127,17 @@ export class AssessmentController {
   @ApiConflictResponse({
     description: 'Assessment with this name and type already exists',
   })
-  update(
+  async update(
     @Param('assessment_id', ParseIntPipe) id: number,
-    @Body() updateAssessmentDto: UpdateAssessmentDto
+    @Body() updateAssessmentDto: UpdateAssessmentDto,
+    @AuthUser() user: User
   ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.update(id, updateAssessmentDto);
   }
 
@@ -137,7 +149,16 @@ export class AssessmentController {
   @Delete(':assessment_id')
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
-  delete(@Param('assessment_id', ParseIntPipe) id: number) {
+  async delete(
+    @Param('assessment_id', ParseIntPipe) id: number,
+    @AuthUser() user: User
+  ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.delete(id);
   }
 
@@ -149,7 +170,16 @@ export class AssessmentController {
   @Post(':assessment_id/complete')
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
-  complete(@Param('assessment_id', ParseIntPipe) id: number) {
+  async complete(
+    @Param('assessment_id', ParseIntPipe) id: number,
+    @AuthUser() user: User
+  ) {
+    const assessment = await this.assessmentService.userInAssessment(id, user);
+
+    if (!assessment) {
+      throw new ForbiddenException();
+    }
+
     return this.assessmentService.complete(id);
   }
 
@@ -160,7 +190,6 @@ export class AssessmentController {
    * @returns Checkpoint saved
    */
   @Post(':assessment_id/save')
-  @UseGuards(AuthGuard('jwt'))
   @ApiOkResponse({ description: 'Checkpoint saved' })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
   async saveCheckpointAnswer(
@@ -187,7 +216,6 @@ export class AssessmentController {
    * @returns Saved checkpoints
    */
   @Get(':assessment_id/save')
-  @UseGuards(AuthGuard('jwt'))
   @ApiResponse({
     description: 'Saved checkpoints',
     type: SaveCheckpointDto,
@@ -216,8 +244,7 @@ export class AssessmentController {
    * @returns updated Assessment object
    */
   @Post(':assessment_id/feedback')
-  @UseGuards(AuthGuard('jwt'), RolesGuard)
-  @Roles(Role.ASSESSOR)
+  @Roles(Role.ASSESSOR, Role.ADMIN)
   @ApiResponse({ description: 'Assessment', type: AssessmentDto })
   @ApiNotFoundResponse({ description: 'Assessment not found' })
   feedback(
@@ -229,7 +256,6 @@ export class AssessmentController {
 
   @Get(':assessment_id/feedback')
   @ApiTags('feedback')
-  @UseGuards(AuthGuard('jwt'))
   @ApiResponse({
     description: 'Recommendations',
     type: RecommendationDto,
